Allow pcn_log to be split in multiple lines
This commit fixes a "bug" in the pcn_log function that did not allow
to split the pcn_log body into multiple lines, for example by putting
the text and the parameters in different lines.
It uses an improved regex function to check the end of the
pcn_log function and replace it with the correct version so that
compilation errors can be avoided.

Fixes: #1 (pcn_log cannot be split in multiple lines)
Signed-off-by: Sebastiano Miano <mianosebastiano@gmail.com>