create transparent cubes
A transparent cube is a new kind of cubes that does not have the concept of
ports, they are instead associated to a cube's port.

They are composed by two sets of eBPF programs, ingress and egress, those
are invoked when there is an incomming/outgoing packet on the nic.

An eBPF program in a transparent cube can pass the packet (modifying it in some
cases like the nat) or drop it.

Signed-off-by: Mauricio Vasquez B <mauriciovasquezbernal@gmail.com>