Add support for bpf_redirect_map in XDP Cubes
This commit introduces the possibility to use the bpf_redirect_map
helper for XDP programs. This helper has improved performance compared
to the normal bpf_redirect function that is used for other programs.
It uses a new map called BPF_DEVMAP that contains a mapping between a virtual
index (used internally by the Cube to indicate a specific port) and the
ifindex of the physical interface connected to the Cube port.
This map can only be used for sending packets out on a netdev (either physical
or virtual) but cannot be used for the Cube-to-Cube communication.
The approach adopted here is to check the value in the forward_chain_ map;
if the value is 0 than there are no entries for that output port and the
devmap is used.

Signed-off-by: Sebastiano Miano <mianosebastiano@gmail.com>