move help generation from services to daemon
The help that is provided to the user was generated by the code generation tools
from the datamodel.  This approach presented some disadvantages, the main one
was the complexity of this, it was difficult to change as this required to
generate the code of the services even with a simple change in the datamodel.

This commit moves the logic to generate the help from the services to the daemon,
help is still generated from datamodel but it is done dynamically by polycubed.

This commit doesn't present any change to the user, the help output is the same
as before, however this makes it a lot easier to improve this when required in
the future.

Signed-off-by: Mauricio Vasquez B <mauriciovasquezbernal@gmail.com>