services: remove static functions
The code structure so far included static functions in the complex elements to
to handle the reading, creation, and deletion of such elements.
This approach make code readibility a little bit difficult as there were many
calls from the parent to the son and back.

This commit removes those static functions, now the handling of that elements is
done directly in the parent.

Signed-off-by: Mauricio Vasquez B <mauriciovasquezbernal@gmail.com>