fix logging in services
after 5a3e91ea90a2 ("implement base datamodel handling in polycubed") the
functionality to change the logging level of a cube was broken.  This because
the logging level was only being changed in the datapath but the control plane
logger (the actual logger that prints it to the terminal) wasn't being updated,
as a result some datapath logs messages were being lost.

This commits creates a new callback that is passed to the framework so when
the logging level is changed the cube is notified and it can update its logger.

Fixes: 5a3e91ea90a2 ("implement base datamodel handling in polycubed")

Signed-off-by: Mauricio Vasquez B <mauriciovasquezbernal@gmail.com>