pcn-k8switch: fix bad initialization logic
Before 09356c4c033 ("services: remove generate_code and generate_code_vector")
the k8switch was created with a stub program and then reloaded with the initial
configuration.  That approach is far from optimal as the code is immediately
reloaded, then 09356c4c033 changed that to created the cube without any eBPF
program and then add it with the initial configuration.

However, the call to reload_program was still there, it fails because the
stub is not present anymore.  This commits fix that by calling add_program
when the cube is created.

Fixes: e09356c4c033 ("services: remove generate_code and generate_code_vector")

Signed-off-by: Mauricio Vasquez B <mauriciovasquezbernal@gmail.com>