polycubed: fix list autocompletion
List autocompletion was partially broken because a look up with the wrong key
name was done.

This commit also adds some logic to avoid suggesting the "add", "del", "show"
commands when in the middle of a list completion. ex:

polycubectl cube_name complex_list key1 <tab>:  suggest list of key2 for
elements that also matchs key1.

Fixes: 95aa7415d741 ("implement bash autocompletion")

Signed-off-by: Mauricio Vasquez B <mauriciovasquezbernal@gmail.com>