add service to base datamodel
There is an ongoing work to implement a /cubes/ api that will allow to create
a list of cubes with a single request.  In order to support that operation
there should be a service field that indicates the service type of the cube.

This commit adds the service to be base datamodel and updates polycubed to
support it.

Signed-off-by: Mauricio Vasquez B <mauriciovasquezbernal@gmail.com>