libpolycube: provide pkg-config
Services that are not shipped with polycube (standalone services) need to
know which libraries and headers files should use in the building process.

This commit adds a pkg-config file for libpolycube, so services can use it for
compiling.

This commit also fixes some include errors to allow such standalone services to
compile.

Signed-off-by: Mauricio Vasquez B <mauriciovasquezbernal@gmail.com>