polycubed: set missing default elements only at initialization time
polycubed fills the default elements on a request if they are missing, this
guarantees that the default elements are always set when the request arrives to
the service implementation, hence the developer has not to take care of it.

When an update (PATCH) request is done those fields should not be set, because
their value will be modified even if the client wanted to update other elements.

Fixes: d8d280cd96a4 ("Added new REST API and validation code")

Reported-by: Elis Lulja <elis.lulja@gmail.com>
Signed-off-by: Mauricio Vasquez B <mauriciovasquezbernal@gmail.com>