polycubed: move logic to detect deleted interface to service controller
When an interface that is connected to a cube's port is deleted, it is needed
to remove the peer from that port.

This commit creates a listener function in the service controller class that is
on charge of removing the peer of a port when the interface is deleted,
this also removes any possible attach to transparent services.

This commit also solves a bug when transparent services where connected to a
interface and that was deleted without detaching first:

create veth1
polycubectl anytransparentservice add t0
polycubectl attach t0 veth1
// traffic is received by t0
polycubetl del t0
delete veth1

create veth1
polycubectl anytransparentservice add t0
polycubectl attach t0 veth1
// at this point t0 does not receives the packets

Signed-off-by: Mauricio Vasquez B <mauriciovasquezbernal@gmail.com>