add topology load at startup and save to memory and file after each modification
This feature allows to load a topology from file at startup, either from the default folder or from a user selected file, like

polycubed --cubes-dump ~/Desktop/cubes.yaml

keeping the whole configuration in memory.
After each modification in any of the topology of the cube, the configuration in memory is updated and another thread saves it to file updating it too.
It is possible to start the daemon with an empty topology by using the flag --cubes-init: if the file contains a topology, at the first update it is overwritten.
It is possible to start the daemon without the dumping to file functionality by using the flag --cubes-nodump.

Signed-off-by: Riccardo Marchi <riccardo.marchi4@gmail.com>