pcn-firewall: implement support for insert/delete
this commit introduces the support for:
- insert: insert a rule at position id
and translate forward all the following rules.
by default inserts in position id=0
that means at the beginning of the ruleset
- delete: delete a rule taking as parameter the
rule fields (not the id). This could help in
scenarios in witch we don't care about indexes.
if multiple rules are present with same fields
then the firts matching is taken

Signed-off-by: Matteo Bertrone <m.bertrone@gmail.com>