polycubed: fix busy wait in netlink socket
select() updates the timeout structure, so if this is not intialized it is
going to be 0 at some point and then create a busy wait loop, using 100% of a
cpu core.

This commit moves the assignent of the timeout structure to its original
location and adds the comment that was removed by mistake.

Fixes: 553afa701868 ("polycubed: use raw socket for netlink")

Signed-off-by: Mauricio Vasquez B <mauriciovasquezbernal@gmail.com>