Add external stp library in service bridge
This C library will manage all the configurations of each STP instance.
Each instance of class Stp will have an instance of struct stp as a
counterpart, and the same will be for PortsStp class.
Therefore, these classes will interact with the library in order to get
the correct configuration of each parameter (state of the ports and so on).

Signed-off-by: Gianluca Scopelliti <gianlu.1033@gmail.com>