polycubed: add namespace class
The namespace class allows to perform different operation with namespaces
as create, delete, open and run code on them.

A namespace object is a handler of an existing namespace, this can be created by:
Namespace::create(): Creates a new namespace and returns an object representing it
Namesapce::open(): Opens an already existing namespace and returns an object for
handling it

Other methods are:
execute(): runs a piece of code in in this namespace namespace.
remove(): removes the namespace
set_id(): sets the netlink id for this

The threads of polycubed cannot be moved to another namespace because it could
cause issues with the injected eBPF programs, so this class has an worker
thread (shared among all instances) that is used to run functions in different
namespaces.
An event queue with a condition variable are used to send tasks to the worker
thread, also a lazy logic is implemented to have the thread only when there are
Namespace instances.

Signed-off-by: Mauricio Vasquez B <mauriciovasquezbernal@gmail.com>