polycubectl: fix reading from standard input
After 074ece5b349a ("polycubectl: add reading from standard input capability")
polycubectl tries to read the standard input in all the invocations, this
causes commands typed by the user to hang up.

This commit uses a logic to detect if the standard input is being piped, in
that case it is used.

This bug was probably not detected for two reasons:
- We concentrate in test the new feature, i.e, loading configurations from files,
without testing the standard behaviour.
- The bug is not present in our CI (jenkins) probably because it executes
polycubect without stdin.

Fixes: 074ece5b349a ("polycubectl: add reading from standard input capability")

Reported-by: Gianluca Scopelliti <gianlu.1033@gmail.com>
Signed-off-by: Mauricio Vasquez B <mauriciovasquezbernal@gmail.com>