Add implementation of service bridge
Features:
- The datapath code is regenerated automatically whenever stp is enabled
or disabled, so it is not necessary to configure that parameter at the
creation of the bridge.
- For each port, there are one or more PortsStp instances according to
the VLANs allowed for that port.
- A Stp instance is created automatically, when the first PortStp instance
for the same VLAN is created.
- In the same way, a Stp instance is destroyed automatically, when the last
PortsStp instance for the same VLAN is destroyed.
- Since the management of Stp/PortsStp instances is done automatically, the
user is not allowed to create/destroy manually those instances, but only to
configure their parameters.
- Containers Access/Trunk do not exist at the same time; if the port is
configured in access mode the user can't configure trunk parameters and
vice versa

Signed-off-by: Gianluca Scopelliti <gianlu.1033@gmail.com>