polycubed: fix packet out deadlock when attaching transparent cubes
A dead lock is possible when a transparent cube is added to a port
and there is a packetout operation at the same time.
If port1 and port2 are connected, a packetout operation is going
on port1 and a cube is attached to port2

Thread1:                          | Thread2:
port1->send_packet_out()          | port2->add_cube()
  lock(port1)                     |   lock(port2)
  ...                             |   ...
  port2->get_index()              |   port1->set_next_index()
    lock(port2)                   |     lock(port1)

The main cause of this is that send_packet_out() uses get_index() that
internally uses a lock.  This commit modifies get_index() so it doesn't
require to lock but returns a variable that is calculated before.

Signed-off-by: Mauricio Vasquez B <mauriciovasquezbernal@gmail.com>