polycubed: set peers after all cubes are created
When a list of cubes is created this is possible that a port in a cube has as
peer another port in a cube that hasn't been created yet, this causes an
exception that aborts the port's creation.

This commit strips the port's peer and sets them after all cubes are created.

Fixes: fbf9bc95246a ("add /cubes/ creation api")
Signed-off-by: Mauricio Vasquez B <mauriciovasquezbernal@gmail.com>