add list of attached cubes to port
So far there is not a way to know what are the transparent cubes attached
to a given port.  This commit extends the base datamodel with a list of cubes
in the ports and adds all the logic to handle them.

This new resource is init-only-config, this commit also handles all the logic
when a port or a cube is created.

There are two special conditions that are handled, having in mind that when
a transparent cube is attached it can query the port for some parameters we
have that:
- if the parent cube is being created, it's is not visible yet.
- if the parent port is being created this is possible that some parameters
are not initialized yet.

The solution for both conditions is to first create the element (cube or port)
and then apply the configuration to attach the transparent cubes.  This approach
would not be necessary once a proper subscrition mechanishm is implemented.

Signed-off-by: Mauricio Vasquez B <mauriciovasquezbernal@gmail.com>