pcn-k8s: Reformat controllers structure
The controllers package is being restructured to be simpler.
The purpose is to prevent the need of passing the single controllers
around the application all the time, but rather use the package itsef.
This way, it works like Kubernetes API: just call the package and
then get the one you need.
Example: pcn_controllers.Pods().Function()
Logger now includes the name of the function by default.

Signed-off-by: Elis Lulja <elis.lulja@gmail.com>