pcn-k8s: Reformat & simplify K8s Policy Controller
Since the controllers package is being reformatted to be the
"single-point" of contact for all the controllers, the Kubernetes
Network Policy Controller undergoes the same process.
It has been stripped of some variables that are now part of the package
itself and, thus, of unnecessary imports.
The controllers.go file is meant to be the only one that can start
the k8s network policy controller, so some function are now
unexported.

Signed-off-by: Elis Lulja <elis.lulja@gmail.com>