Dynmon enhancement - Key export feature
This commit addresses the 2nd point discussed in polycube-network#291.
It has been introduced the key extraction, meaning that if the bpf map is a
key-value type one, not only the value is exported, but also the associated key.

Still, for map-in-map types there is not full support yet, due to many problem related to
the map info retrieval.

Signed-off-by: Simone Magnani <simonemagnani.96@gmail.com>