Dynmon feature Empty map on read
This commit introduces the "empty man when read" feature. Accordingly to its type,
the map is erased as specified by the user when this feature is enabled.
If it is an Hash map then the entry is completely removed, otherwise it is set
to zero (cannot remove an entry from a static array, se empty it)

Signed-off-by: Simone Magnani <simonemagnani.96@gmail.com>