Dynmon Compiler for Swappable code
This commit include a major improvement in Dynmon lifecycle: a runtime compiler for Swappable code!
Its aim is to compile the code to achieve user required features like MAP-SWAP, trying to minimize the
execution swap overhead. Firstly, it try to use a "change pivoting map index" techniques; if that fails, it
uses the fallback "replace the entire code" one, which makes every GET_METRIC slower (5ms vs 400ms) but at
least the functionality is as required by the user

Signed-off-by: Simone Magnani <simonemagnani.96@gmail.com>