firewall batch operations + batch update + changed default action policy + removed "transactional" + docs and tests (#323)
This commit aims to introduce batch operation in firewall, to optimize network usage under certain
situations. For example, it would be both time and resource consuming performing 1000 requests to add/remove
rules, which was still required in the previous "transaction" mode. So I decided to introduce an endpoint which can accept a list of operations to be performed.

Moreover, I have also changed the default action policy, previously set to DROP. I set it to FORWARD, to avoid that
a remote client which interacts with Polycube is completely cut out from the system, since the DROP policy would
block ALL the incoming traffic.

Signed-off-by: Simone Magnani <simonemagnani.96@gmail.com>