Fix map reuse when reloading tc egress program of xdp cubes
When reloading a program the new program needs to reuse the maps of
the old one. In the EGRESS path of XDP cubes moreover the two
programs (TC and XDP) need to share the same maps to beahve as
one. For the TC program stealing the map from the old TC program
didn't work, causing it to use a new map different from the XDP
counterpart. With this fix the TC program directly uses the map of
the XDP counterpart.