Fall back to map lookup when cube's redirection code becomes too big
The code of the pcn_pkt_redirect() function of standard cubes is
dynamically generated to handle redirection to other cubes/interfaces
with a switch-case statement. When there are too many peers however
(> 17) this code becomes too big and is rejected by the verfier.
This commit keeps the optimized path for the first peers and adds a
fall back table to redirect to other peers.